requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
